requirements.txt
async-lru==1.0.2
pyaes==1.6.1
pyasn1==0.4.8
Pyrogram==1.2.11
PySocks==1.7.1
python-dotenv==0.18.0
rsa==4.7.2
Telethon==1.24.0
TgCrypto==1.2.2
